// src/Models/home.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';
import { Article } from './article';
import { HeroUnit } from './hero_unit';
import { Cafe } from './cafe';
import { FactAboutUs } from './fact_about_us';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type Home = IContentItem<{
  metadataOgDescription: Elements.TextElement;
  metadataMetaTitle: Elements.TextElement;
  metadataOgTitle: Elements.TextElement;
  articles: Elements.LinkedItemsElement<Article>;
  heroUnit: Elements.LinkedItemsElement<HeroUnit>;
  metadataMetaDescription: Elements.TextElement;
  metadataTwitterSite: Elements.TextElement;
  ourStory: Elements.LinkedItemsElement<FactAboutUs>;
  cafes: Elements.LinkedItemsElement<Cafe>;
  metadataTwitterImage: Elements.AssetsElement;
  metadataTwitterCreator: Elements.TextElement;
  metadataTwitterTitle: Elements.TextElement;
  metadataTwitterDescription: Elements.TextElement;
  metadataOgImage: Elements.AssetsElement;
  urlPattern: Elements.UrlSlugElement;
  sitemap: Elements.TaxonomyElement;
  contact: Elements.RichTextElement;
}>;